feature(): move Response class to response dir
